Tidied webpack configuration
