+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "flux": "^3.1.3",
     "history": "^1.17.0",
+    "immutable": "^3.8.2",
     "react": "^16.4.1",
     "react-bootstrap": "^0.32.1",
     "react-dom": "^16.4.1",
@@ -31,6 +32,7 @@
     "style-loader": "^0.22.1",
     "webpack": "^4.16.5",
     "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.5"
+    "webpack-dev-server": "^3.1.5",
+    "webpack-merge": "^4.1.4"
   }
 }
