+++ webpack.config.js
@@ -1,28 +1,17 @@
 'use strict';
 const webpack = require('webpack');
 const path = require('path');
+const webpackMerge = require('webpack-merge');
+const common = require('./webpack.config.common.js');
 
 //const BUILD_DIR = path.resolve(__dirname, 'frontend/public/js');
 //Let's build to public so that Lumen can serve it.
 const BUILD_DIR = path.resolve(__dirname, 'public/js');
-const APP_DIR = path.resolve(__dirname, 'frontend/src');
 
 const config = {
     mode: 'development',
-    context: APP_DIR,
-    entry: './index.js',
     output: {
-        path: BUILD_DIR,
-        filename: 'bundle.js',
-        publicPath: "/"
-    },
-    //Add path aliases here for easier access in frontend imports
-    resolve:{
-        alias:{
-            '@Routes' : path.resolve(__dirname,'frontend/src/routes'),
-            '@Components' : path.resolve(__dirname,'frontend/src/components'),
-            '@Services' : path.resolve(__dirname,'frontend/src/services'),
-        }
+        path: BUILD_DIR
     },
     module: {
         rules: [
@@ -31,17 +20,7 @@ const config = {
                 exclude: /node_modules/,
 		loader: 'babel-loader',
 		query: { presets: [ 'es2015', 'react' ] }
-            },
-	    {
-		test: /\.(png|jpg)$/,
-		use: [{
-		    loader: 'url-loader'
-		}]
-	    },
-	    {
-		test: /\.css$/,
-		use: ["style-loader", "css-loader"]
-	    }
+            }
         ]
     },
     devServer: {
@@ -51,5 +30,5 @@ const config = {
     }
 };
 
-module.exports = config;
+module.exports = webpackMerge(common,config);
